Rename sqlrepo to db
