+++ backend/app/adapter/db/userurl.go
@@ -1,9 +1,9 @@
-package sqlrepo
+package db
 
 import (
 	"database/sql"
 	"fmt"
-	"short/app/adapter/sqlrepo/table"
+	"short/app/adapter/db/table"
 	"short/app/usecase/repo"
 )
